channeldb: move fwding policy code to new file
